refactor: split internal/editor/fseditor.go by operation kind
544 lines holding the full-screen editor's construction, key dispatch, text
edits and cursor movement, split the way the backlog proposed:

  fseditor.go         196  construction, caller settings, Run loop, redraw
  fseditor_edit.go    130  operations that change the buffer
  fseditor_keys.go    119  key and command dispatch
  fseditor_cursor.go  115  cursor movement and scroll-to-cursor

The split line is whether an operation changes the buffer: cursor movement
and scrolling do not, edits do, and dispatch decides which to call.

Pure move: no declaration body changed. Verified with a go/parser pairwise
diff of every top-level declaration (34 old, 34 new, 0 missing, 0 mismatched,
0 extra) and a comment multiset diff showing nothing removed. No test file
touched.
