+++ internal/editor/fseditor.go
@@ -7,6 +7,10 @@ import (
 	"github.com/gliderlabs/ssh"
 )
 
+// The full-screen message editor: construction, the settings a caller supplies
+// before it runs, and the Run loop itself. Key dispatch, text edits and cursor
+// movement live in the fseditor_* files beside this one.
+
 // FSEditor is the full-screen message editor
 type FSEditor struct {
 	// Components
@@ -180,361 +184,9 @@ func (e *FSEditor) Run() (string, bool, error) {
 	return content, e.saved, nil
 }
 
-// handleKey processes a single key press
-func (e *FSEditor) handleKey(key int) {
-	// Handle key based on type
-	switch key {
-	case KeyEnter:
-		e.handleReturn()
-	case KeyBackspace:
-		e.handleBackspace()
-	case KeyTab:
-		e.handleTab()
-
-	// Escape: show option lightbar menu (Save / Abort / Edit / Help / Quote)
-	case KeyEsc:
-		cmdType := e.commands.ShowEscapeMenu(e.input)
-		if cmdType != CommandNone {
-			e.handleCommand(cmdType)
-		}
-
-	// Editor commands (shown in footer: CTRL (A)Abort (Z)Save (Q)Quote)
-	case KeyCtrlA: // Abort
-		e.handleCommand(CommandAbort)
-	case KeyCtrlZ: // Save
-		e.handleCommand(CommandSave)
-	case KeyCtrlQ: // Quote
-		e.handleCommand(CommandQuote)
-
-	// Navigation
-	case KeyCtrlE: // Up
-		e.moveCursorUp()
-	case KeyCtrlX: // Down
-		e.moveCursorDown()
-	case KeyCtrlS: // Left
-		e.moveCursorLeft()
-	case KeyCtrlD: // Right
-		e.moveCursorRight()
-	case KeyCtrlW: // Home
-		e.moveCursorHome()
-	case KeyCtrlP: // End
-		e.moveCursorEnd()
-	case KeyCtrlR: // Page Up
-		e.pageUp()
-	case KeyCtrlC: // Page Down (note: normally quit in terminals, but remapped here)
-		e.pageDown()
-	case KeyCtrlF: // Word Right
-		e.moveCursorWordRight()
-
-	// Edit commands
-	case KeyCtrlV: // Toggle insert/overwrite
-		e.toggleInsertMode()
-	case KeyCtrlG: // Delete character at cursor
-		e.handleDeleteKey()
-	case KeyCtrlT: // Delete word
-		e.deleteWord()
-	case KeyCtrlY: // Delete line
-		e.deleteLine()
-	case KeyCtrlJ: // Join lines
-		e.joinLines()
-	case KeyCtrlN: // Split line
-		e.splitLine()
-	case KeyCtrlB: // Reformat paragraph
-		e.reformatParagraph()
-	case KeyCtrlL: // Redraw screen
-		e.redrawScreen()
-
-	default:
-		// Check if it's a printable character
-		if IsPrintable(key) {
-			e.insertCharacter(rune(key))
-		}
-	}
-}
-
-// handleCommand processes editor commands (CTRL-A/Z/Q and help/view)
-func (e *FSEditor) handleCommand(cmdType CommandType) {
-	switch cmdType {
-	case CommandSave:
-		if e.commands.HandleSave() {
-			// Show "Saving..." in the prompt row before exiting
-			e.screen.GoXY(1, e.screen.PromptRow())
-			e.screen.ClearEOL()
-			e.screen.WriteDirectProcessed("|15Saving...")
-			e.saved = true
-			e.quit = true
-		} else {
-			// Error message already written; wait for key then restore footer
-			_, _ = e.input.ReadKey() // wait for any key
-			e.screen.DisplayFooter()
-			e.screen.RefreshScreen(e.buffer, e.topLine, e.currentLine, e.currentCol, e.insertMode, true)
-		}
-
-	case CommandAbort:
-		if e.commands.HandleAbort(e.input) {
-			e.saved = false
-			e.quit = true
-		} else {
-			e.screen.RefreshScreen(e.buffer, e.topLine, e.currentLine, e.currentCol, e.insertMode, true)
-		}
-
-	case CommandQuote:
-		line, col := e.commands.HandleQuote(e.input, e.currentLine, e.currentCol)
-		e.currentLine = line
-		e.currentCol = col
-		e.modified = true
-		e.screen.FullRedraw(e.buffer, e.topLine, e.currentLine, e.currentCol, e.insertMode)
-
-	case CommandHelp:
-		e.commands.HandleHelp(e.input)
-		e.screen.FullRedraw(e.buffer, e.topLine, e.currentLine, e.currentCol, e.insertMode)
-
-	case CommandView:
-		e.commands.HandleView(e.input)
-		e.screen.FullRedraw(e.buffer, e.topLine, e.currentLine, e.currentCol, e.insertMode)
-	}
-}
-
-// insertCharacter inserts a printable character at the cursor
-func (e *FSEditor) insertCharacter(ch rune) {
-	if e.insertMode {
-		e.buffer.InsertChar(e.currentLine, e.currentCol, ch)
-	} else {
-		e.buffer.OverwriteChar(e.currentLine, e.currentCol, ch)
-	}
-
-	e.currentCol++
-	e.modified = true
-
-	// Check for word wrap
-	newLine, newCol := e.wordWrapper.WrapAfterInsert(e.currentLine, e.currentCol)
-	e.currentLine = newLine
-	e.currentCol = newCol
-}
-
-// handleReturn processes the Enter key
-func (e *FSEditor) handleReturn() {
-	// Split line at cursor position
-	if e.buffer.SplitLine(e.currentLine, e.currentCol) {
-		// Mark the split line as a hard newline (user-created break)
-		e.buffer.SetHardNewline(e.currentLine, true)
-		e.currentLine++
-		e.currentCol = 1
-		e.modified = true
-	}
-}
-
-// handleBackspace processes the Backspace key
-func (e *FSEditor) handleBackspace() {
-	newLine, newCol, changed := e.wordWrapper.HandleBackspace(e.currentLine, e.currentCol)
-	if changed {
-		e.currentLine = newLine
-		e.currentCol = newCol
-		e.modified = true
-	}
-}
-
-// handleDeleteKey processes the Delete key
-func (e *FSEditor) handleDeleteKey() {
-	newLine, newCol, changed := e.wordWrapper.HandleDelete(e.currentLine, e.currentCol)
-	if changed {
-		e.currentLine = newLine
-		e.currentCol = newCol
-		e.modified = true
-	}
-}
-
-// handleTab inserts tab spaces
-func (e *FSEditor) handleTab() {
-	// Insert 4 spaces for tab
-	for i := 0; i < 4; i++ {
-		e.insertCharacter(' ')
-	}
-}
-
-// moveCursorUp moves cursor up one line
-func (e *FSEditor) moveCursorUp() {
-	if e.currentLine > 1 {
-		e.currentLine--
-		// Adjust column if line is shorter
-		lineLen := e.buffer.GetLineLength(e.currentLine)
-		if e.currentCol > lineLen+1 {
-			e.currentCol = lineLen + 1
-		}
-	}
-}
-
-// moveCursorDown moves cursor down one line
-func (e *FSEditor) moveCursorDown() {
-	lineCount := e.buffer.GetLineCount()
-	if e.currentLine < lineCount {
-		e.currentLine++
-		// Adjust column if line is shorter
-		lineLen := e.buffer.GetLineLength(e.currentLine)
-		if e.currentCol > lineLen+1 {
-			e.currentCol = lineLen + 1
-		}
-	}
-}
-
-// moveCursorLeft moves cursor left one character
-func (e *FSEditor) moveCursorLeft() {
-	if e.currentCol > 1 {
-		e.currentCol--
-	} else if e.currentLine > 1 {
-		// Move to end of previous line
-		e.currentLine--
-		e.currentCol = e.buffer.GetLineLength(e.currentLine) + 1
-	}
-}
-
-// moveCursorRight moves cursor right one character
-func (e *FSEditor) moveCursorRight() {
-	lineLen := e.buffer.GetLineLength(e.currentLine)
-	if e.currentCol <= lineLen {
-		e.currentCol++
-	} else if e.currentLine < e.buffer.GetLineCount() {
-		// Move to start of next line
-		e.currentLine++
-		e.currentCol = 1
-	}
-}
-
-// moveCursorHome moves cursor to start of line
-func (e *FSEditor) moveCursorHome() {
-	e.currentCol = 1
-}
-
-// moveCursorEnd moves cursor to end of line
-func (e *FSEditor) moveCursorEnd() {
-	lineLen := e.buffer.GetLineLength(e.currentLine)
-	e.currentCol = lineLen + 1
-}
-
-// moveCursorWordRight moves cursor to start of next word
-func (e *FSEditor) moveCursorWordRight() {
-	e.currentCol = e.wordWrapper.FindWordRight(e.currentLine, e.currentCol)
-}
-
-// pageUp scrolls up one page
-func (e *FSEditor) pageUp() {
-	scrollAmount := e.screen.GetScreenLines() - 1
-	e.currentLine -= scrollAmount
-	if e.currentLine < 1 {
-		e.currentLine = 1
-	}
-}
-
-// pageDown scrolls down one page
-func (e *FSEditor) pageDown() {
-	scrollAmount := e.screen.GetScreenLines() - 1
-	lineCount := e.buffer.GetLineCount()
-	e.currentLine += scrollAmount
-	if e.currentLine > lineCount {
-		e.currentLine = lineCount
-	}
-	if e.currentLine < 1 {
-		e.currentLine = 1
-	}
-}
-
-// toggleInsertMode toggles between insert and overwrite modes
-func (e *FSEditor) toggleInsertMode() {
-	e.insertMode = !e.insertMode
-}
-
-// deleteWord deletes the word to the right of the cursor
-func (e *FSEditor) deleteWord() {
-	newLine, newCol, changed := e.wordWrapper.DeleteWord(e.currentLine, e.currentCol)
-	if changed {
-		e.currentLine = newLine
-		e.currentCol = newCol
-		e.modified = true
-	}
-}
-
-// deleteLine deletes the current line
-func (e *FSEditor) deleteLine() {
-	if e.buffer.DeleteLine(e.currentLine) {
-		e.modified = true
-		// Ensure cursor is on a valid line
-		if e.currentLine > e.buffer.GetLineCount() {
-			e.currentLine = e.buffer.GetLineCount()
-			if e.currentLine < 1 {
-				e.currentLine = 1
-			}
-		}
-		e.currentCol = 1
-	}
-}
-
-// joinLines joins current line with next line and reflows
-func (e *FSEditor) joinLines() {
-	if e.currentLine >= e.buffer.GetLineCount() {
-		return
-	}
-	// Clear hardNewline so reflow can flow across the boundary
-	e.buffer.SetHardNewline(e.currentLine, false)
-	if e.buffer.JoinLines(e.currentLine) {
-		e.modified = true
-		newLine, newCol := e.wordWrapper.ReflowRange(e.currentLine, e.currentLine, e.currentCol)
-		e.currentLine = newLine
-		e.currentCol = newCol
-	}
-}
-
-// splitLine splits the current line at the cursor (user-initiated Ctrl+N)
-func (e *FSEditor) splitLine() {
-	if e.buffer.SplitLine(e.currentLine, e.currentCol) {
-		// Mark the split line as a hard newline (user-created break)
-		e.buffer.SetHardNewline(e.currentLine, true)
-		e.currentLine++
-		e.currentCol = 1
-		e.modified = true
-	}
-}
-
-// reformatParagraph reformats the current paragraph
-func (e *FSEditor) reformatParagraph() {
-	lastLine := e.wordWrapper.ReformatParagraph(e.currentLine)
-	e.modified = true
-	// Keep cursor on a valid line
-	if e.currentLine > lastLine {
-		e.currentLine = lastLine
-	}
-	e.currentCol = 1
-	// Force full redraw
-	e.screen.FullRedraw(e.buffer, e.topLine, e.currentLine, e.currentCol, e.insertMode)
-}
-
 // redrawScreen forces a complete screen redraw
 func (e *FSEditor) redrawScreen() {
 	e.screen.FullRedraw(e.buffer, e.topLine, e.currentLine, e.currentCol, e.insertMode)
-}
-
-// ensureCursorVisible adjusts the view to keep the cursor visible
-func (e *FSEditor) ensureCursorVisible() {
-	screenLines := e.screen.GetScreenLines()
-	oldTopLine := e.topLine
-
-	// Check if cursor is above visible area
-	if e.currentLine < e.topLine {
-		e.topLine = e.currentLine
-	}
-
-	// Check if cursor is below visible area
-	if e.currentLine >= e.topLine+screenLines {
-		e.topLine = e.currentLine - screenLines + 1
-		if e.topLine < 1 {
-			e.topLine = 1
-		}
-	}
-
-	// If topLine changed (scrolling occurred), clear screen cache to force redraw
-	if e.topLine != oldTopLine {
-		e.screen.ClearCache()
-		e.lastTopLine = e.topLine
-	}
 }
 
 // HandleResize handles terminal resize events
